requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
idna==3.10
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pydentic==0.0.1.dev3
python-dotenv==1.0.1
python-stdnum==1.20
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
